requirements.txt
streamlit
openai
langchain
autogen
pytest
pytest-timeout
python-dotenv
requests
httpx
pydantic
rich 